refactor: ApiDoc을 ItdocBuilderEntry로 이름 변경
ApiDoc이라는게 에매한 이름 같아서 좀 더 명확한 의미를 가진 ItdocBuilderEntry로 변경.
(itdoc를 호출할때 인자로 사용하는 빌더의 진입점이라는 뜻)
