+++ lib/dsl/index.ts
@@ -34,4 +34,4 @@
 export { HttpMethod } from "./enums/HttpMethod"
 export { HttpStatus } from "./enums/HttpStatus"
 export { describeAPI, itDoc, field } from "./interface"
-export type { ApiDocOptions } from "./interface/ApiDoc"
+export type { ApiDocOptions } from "./interface/ItdocBuilderEntry"
